Refactor Splash component
