+++ src/app/components/App/components/Splash/components/ErrorMessage.js
@@ -1,0 +1,9 @@
+const ErrorMessage = () => (
+  <div className="ui error message">
+    <ui className="list">
+      <li>Please enter your name.</li>
+    </ui>
+  </div>
+)
+
+export default ErrorMessage
