+++ src/app/components/App/components/Splash/components/Header.js
@@ -1,0 +1,11 @@
+const Header = () => (
+  <h2 className="ui center aligned icon header">
+    <i className="youtube play icon"></i>
+    <div className="content">
+      Watch YouTube Together!
+      <div className="sub header">Greetings! Say hi to everyone. 🙋</div>
+    </div>
+  </h2>
+)
+
+export default Header
